+++ dorado/cli/duplex.cpp
@@ -553,8 +553,9 @@ int duplex(int argc, char* argv[]) {
         tracker.update_progress_bar(final_stats);
 
         // Report progress during output file finalisation.
-        hts_file.finalise(
-                [&](size_t progress) { tracker.update_post_processing_progress(progress); });
+        hts_file.finalise([&](size_t progress) {
+            tracker.update_post_processing_progress(static_cast<float>(progress));
+        });
 
         tracker.summarize();
         if (!dump_stats_file.empty()) {
